+++ src/Point.cpp
@@ -61,14 +61,14 @@ void Point::putField(String name, String value) {
 
 String Point::toLineProtocol(String includeTags) const {
     String line =  _measurement;
-    line.reserve(1 + _tags.length() + 1 + includeTags.length() + 1 + _fields.length() + 1 + _timestamp.length());
-    if(hasTags()) {
-        line += ",";
-        line += _tags;
-    }
+    line.reserve(1 + includeTags.length() + 1 + _tags.length() + 1 + _fields.length() + 1 + _timestamp.length());
     if(includeTags.length()>0) {
         line += ",";
         line += includeTags;
+    }
+    if(hasTags()) {
+        line += ",";
+        line += _tags;
     }
     if(hasFields()) {
         line += " ";